refactor: moved functions to TodoList
