+++ src/components/TaskList/TaskList.tsx
@@ -1,6 +1,5 @@
 import TaskItem from '../TaskItem/TaskItem.js'
-import { changeDataTask, deleteTask } from '../../api/http.js'
-import { FC, useState } from 'react'
+import { FC } from 'react'
 import { TasksType } from '../../types/todolist.ts'
 import { List } from 'antd'
 import classes from './TaskList.module.css'
@@ -8,52 +7,23 @@ import classes from './TaskList.module.css'
 type TaskListProps = {
     tasks: TasksType[]
     fetchTasksByFilter: () => void
+    handleDeleteTask: (id: number) => void
+    handleChangeDataTask: (id: number, titleTask: string, isDone: boolean) => void
 }
 
-const TaskList: FC<TaskListProps> = ({ fetchTasksByFilter, tasks }) => {
-    const [isFetching, setIsFetching] = useState(false)
-
-    const handleDeleteTask = async (id: number) => {
-        try {
-            setIsFetching(true)
-            await deleteTask(id)
-            await fetchTasksByFilter()
-        } catch {
-            setAndAlertError('Ошибка удаления задачи!')
-        } finally {
-            setIsFetching(false)
-        }
-    }
-
-    const handleChangeDataTask = async (id: number, titleTask: string, isDone: boolean) => {
-        try {
-            setIsFetching(true)
-            await changeDataTask(id, titleTask, isDone)
-            await fetchTasksByFilter()
-        } catch {
-            setAndAlertError('Ошибка изменения задачи!')
-        } finally {
-            setIsFetching(false)
-        }
-    }
-
-    const setAndAlertError = (error: string) => {
-        alert(error)
-    }
-
+const TaskList: FC<TaskListProps> = ({ tasks, handleDeleteTask, handleChangeDataTask }) => {
     return (
         <List className={classes.tasklist}>
-            {!isFetching &&
-                tasks.map(task => (
-                    <TaskItem
-                        key={task.id}
-                        id={task.id}
-                        titleTask={task.title}
-                        isDone={task.isDone}
-                        onChangeData={handleChangeDataTask}
-                        onDelete={handleDeleteTask}
-                    />
-                ))}
+            {tasks.map(task => (
+                <TaskItem
+                    key={task.id}
+                    id={task.id}
+                    titleTask={task.title}
+                    isDone={task.isDone}
+                    onChangeData={handleChangeDataTask}
+                    onDelete={handleDeleteTask}
+                />
+            ))}
         </List>
     )
 }
